snap packaging: uprev dotnet-sdk, move to core20, and various fixes
- uprev .NET SDK to 3.1.402

- migrate snap base to core20

- enable for all published architectures (amd64, arm64, armhf)

Snapcraft supports "advanced grammar" to selectively configure some
aspects of the snapcraft.yaml based on the target architecture.
Unfortunately this does not currently support `source-checksum`
property to fully verify the fetched sources.  This aspect is on the
Snapcraft road to address, but may not be an issue for the real build
infrastructure.

- manually patch 3.1 SDK dotnet rpath & interpreter to the correct paths

Snapcraft automatically patches the rpath and interpreter for all executables
found within a snap.  However, this impacts the `apphost` template binaries,
resulting in some unexpected behaviors and crashes, e.g.:
`*** stack smashing detected ***: terminated`

Disable Snapcraft's automatic patching using the 'no-patchelf' build-attribute,
but then patch the `dotnet` binary to ensure it works properly as a classic
app.

- test scripts used on Ubuntu 16.04, 18.04, 20.04, and Fedora 32:

https://github.com/cjp256/dotnet-sdk-tests

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>